+++ src/router.cc
@@ -1,4 +1,4 @@
-#include "router.h"
+#include "djangocpp/router.h"
 
 void Router::add_route(const string &method, const string &path, Handler handler)
 {
